Rename reader to make clear it is meant to be used for by-date processing (by-batch coming soon)
