+++ tests/op_datasets/etl/ingestion/test_reader.py
@@ -3,7 +3,7 @@ import datetime
 import polars as pl
 
 from op_analytics.coreutils.partitioned.location import DataLocation
-from op_analytics.datapipeline.etl.ingestion.reader import are_inputs_ready
+from op_analytics.datapipeline.etl.ingestion.reader_bydate import are_inputs_ready
 
 MARKER_PATHS_DATA = [
     {
